fcoverage delta array is f64 but halves tile size to compensate

+++ src/shared/coverage.rs
@@ -66,7 +66,7 @@ enum Stage {
 #[derive(Debug, Clone)]
 pub struct Coverage {
     length: u32,                // Total sequence length in bases (e.g., chrom_len)
-    delta: Vec<f32>,            // +w at start, -w at end, length = length + 1 (last is sentinel)
+    delta: Vec<f64>,            // +w at start, -w at end, length = length + 1 (last is sentinel)
     coverage: Option<Vec<f32>>, // Per-base coverage after finalize_coverage, length = length
     bl_mask: Option<Vec<u8>>, // Per-base blacklist mask after finalize_blacklist_prefix, 1 = blacklisted
 
@@ -106,7 +106,7 @@ impl Coverage {
     pub fn new(length: u32) -> Self {
         Self {
             length,
-            delta: vec![0.0; length as usize + 1],
+            delta: vec![0.0_f64; length as usize + 1],
             coverage: None,
             bl_mask: None,
             psum_all: None,
@@ -136,7 +136,7 @@ impl Coverage {
     /// - weight:
     ///     Weight to add, must be finite and >= 0.
     #[inline]
-    pub fn add_fragment_weighted(&mut self, frag: Fragment, weight: f32) -> Result<()> {
+    pub fn add_fragment_weighted(&mut self, frag: Fragment, weight: f64) -> Result<()> {
         if !self.prefix_available() {
             anyhow::bail!(
                 "prefix was dropped; cannot add fragments. Rebuild or create a new Coverage"
@@ -190,7 +190,7 @@ impl Coverage {
     pub fn add_fragment_with_segments(
         &mut self,
         frag: FragmentWithSegments,
-        weight: f32,
+        weight: f64,
     ) -> anyhow::Result<()> {
         if !self.prefix_available() {
             anyhow::bail!(
@@ -256,7 +256,7 @@ impl Coverage {
         // Cumulative sum over delta
         let mut run = 0.0_f64;
         for i in 0..=self.length as usize {
-            run += self.delta[i] as f64;
+            run += self.delta[i];
             if i < self.length as usize {
                 cov[i] = run as f32;
             }
